Update evaluations script to identify checkmates and tactics
Note: the tactics identification is dubious but seems to work in most cases.
It identifies positions where the previous eval to the current eval swung
more than 300 centipawns in either direction. However, that could mean that
there WAS a tactic on the board, but now there's not. Either way, the idea
was to identify positions with tactical complications present and it's doing
a good job of that.